change the path for file uploads to depend on the file being edited.
