+++ src/smc-util/misc2.ts
@@ -640,3 +640,14 @@ export function history_path(path: string): string {
 export function tuple<T extends string[]>(o: T) {
   return o;
 }
+
+
+export function aux_file(path: string, ext: string): string {
+  const s = path_split(path);
+  s.tail += "." + ext;
+  if (s.head) {
+    return s.head + "/." + s.tail;
+  } else {
+    return "." + s.tail;
+  }
+}
